requirements.txt
aiosqlite
telethon
python-dotenv
flask